Rework Faraday probe
This commit reworks the Faraday probe to avoid a few issues. Previously
the probe has set a custom builder as the default connection options to
include the tracing middleware. This resulted in a few issues as the
default block configuration with Faraday did not longer work, e.g.:

  Faraday.new do |f|
    f.use CustomMiddleware
    f.adapter ...
  end

Due to Faraday not using the block at all if a `:builder` option is
given no customization ever happended anymore. The implementation looks
like this:

  @builder = options.builder || begin
    options.new_builder(block_given? ? proc { |b| } : nil)
  end

Options do include the default connections options.

This broke a few gems using Faraday for HTTP communication as no
middlewares were configured anymore.

The new probe patches into the RackBuilder itself injecting the
mnemosyne middleware when finalizing the stack. This will work for all
Faraday connections building their stack *after* the patch has been
applied. Stacks openeded before the probe is loaded will not include
tracing. This is especially important when using the default connection.
If any code finalizes the stack, e.g. by doing a request with
`Faraday#get`, all requests using the default connection will not be
traced.

It is possible to manually add the mnemosyne middleware like this:

  Faraday.new do |f|
    f.use :mnemosyne
  end

The patch will not add another mnemosyne middleware.

The patch itself hooks the `#lock!` method of Faradays `RackBuilder` to
inject the tracing middleware at the last possible chance. It tries to
add it right in front of the adapter to include all changed applied by
previous middlewares. If not adapter is found (e.g. Faraday 1.0) it is
inserted at last place.