feat: Increase compatibility of the require patch
The previous patch did not declare the overwritten require method as
module_function. This led to subtle changes when calling
respond_to?(:require) on object later, which could return true when it
should not.

This commit changes the code to use module_function. It further moves
the loader into a module that is prepended, to avoid the addition of
method aliases.